orchestrator/pyproject.toml
[tool.poetry]
name = "neurohelix-orchestrator"
version = "0.1.0"
description = "Python-based orchestrator for NeuroHelix daily research pipeline"
authors = ["NeuroHelix Team"]
readme = "README.md"
packages = [{include = "nh_cli"}, {include = "services"}, {include = "config"}, {include = "adapters"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "0.20.0"
pydantic = "2.12.4"
rich = "^13.7.0"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
nh = "nh_cli.main:app"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=. --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
